% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_overlap.r
\name{calculate_overlap}
\alias{calculate_overlap}
\title{Calculates the overlap of two stochastic character state distributions. Used internally by correlate.}
\usage{
calculate_overlap(c1_min = 0, c1_max = 0, c2_min = 0, c2_max = 0)
}
\arguments{
\item{c1_min}{Starting point (along branch) of character state 1.}

\item{c1_max}{End point (along branch) of character state 1.}

\item{c2_min}{Starting point (along branch) of character state 2.}

\item{c2_max}{End point (along branch) of character state 2.}
}
\description{
Calculates the overlap of two stochastic character state distributions. Used internally by correlate.
}
\examples{
calculate_overlap(0.1, 0.5, 0.005, 0.4)
}
