% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conditional.R
\name{get_conditional}
\alias{get_conditional}
\title{Calculate conditional probability of character co-occurrence}
\usage{
get_conditional(matrix, probs)
}
\arguments{
\item{matrix}{matrix with intersect probabilities obtained from the function get_intersect.}

\item{probs}{dataframe with marginal probabilities obtained from the function get_marginal.}
}
\description{
This function calculates the conditional probability distribution of observing two character states along the provided phylogenetic tree(s).
}
\examples{
get_conditional(matrix, probs)
}
