% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intersect.R
\name{get_intersect}
\alias{get_intersect}
\title{Calculating character overlap for two independent stochastic character mappings}
\usage{
get_intersect(ntrees = 1, nmaps = 1, smap1 = simmap,
  smap2 = simmap_multi, chars2 = c(0, 1), chars1 = c(0, 1))
}
\arguments{
\item{ntrees}{The number of trees which should be analysed. default = 1}

\item{nmaps}{The number of stochastic character maps which should be analysed per tree. default = 1}

\item{smap1}{List of multiPhylo Objects containing stochastic maps for first character. Created with phytools.}

\item{smap2}{List of multiPhylo Objects containing stochastic maps for second character. Created with phytools.}

\item{chars2}{Vector containing character state names which should be analysed for the first set of characters.}

\item{chars1}{Vector containing character state names which should be analysed for the second set of characters.}
}
\description{
This function calculates the overlapping parts of two (or more) sets of stochastically mapped categorical characters. As input it takes two sets of stochastic maps created with the R package phytools. It will return the proportion of the tree where two characters overlap.
}
\examples{
get_intersect(ntrees = 1, nmaps = 1, smap1 = simmap, smap2 = simmap_multi, chars2 = c(0, 1), chars1 = c(0, 1))
}
