% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marginal.r
\name{get_marginal}
\alias{get_marginal}
\title{Calculate the marginal probability of character occurrence.}
\usage{
get_marginal(sm)
}
\arguments{
\item{smap}{List of multiPhylo objects containing stochastic character maps created with phytools.}
}
\description{
This function calculates the marginal probability of observing a particular character on the provided phylogenetic tree(s). Will return a dataframe containing the marginal probabilities of character occurrence for each character.
}
\examples{
get_marginal(smap)
}
