% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagel_multi.R
\name{pagel_multi}
\alias{pagel_multi}
\title{Runs Pagel's correlation method for multiple trees.}
\usage{
pagel_multi(trees, a, b)
}
\arguments{
\item{a}{Character distribution for first character.}

\item{b}{Character distribution for second character.}

\item{tree}{Multiphylo object with phylogenetic trees.}
}
\description{
Runs Pagel's test for binary character correlation for multiple trees. Borrows a few ideas and lines of code from the phytools function fitPagel.
}
\examples{
pagel_multi(trees, a, b)
}
