% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{plot_corr}
\alias{plot_corr}
\title{Create a correlation matrix}
\usage{
plot_corr(
  x,
  corr = "spearman",
  corr_round = 2,
  corr_col_scale = c("red", "green"),
  order = NULL,
  upper = "scatter",
  labels = NULL,
  label_text_size = 2,
  label_text_font = 1,
  label_text_col = "black",
  corr_text_size = 2,
  corr_text_font = 1,
  corr_text_col = "black",
  point_size = 1,
  point_col = "black",
  point_type = 16,
  title = "Correlation Matrix",
  title_text_size = 1.5,
  title_text_col = "black",
  title_text_font = 2,
  hist_col = "blue",
  hist_border_col = "black",
  hist_border_type = 1,
  hist_border_width = 1,
  density_line_col = "black",
  density_line_type = 1,
  density_line_width = 1,
  trend_line_type = 1,
  trend_line_col = "red",
  trend_line_width = 1,
  margins = 0.5,
  frame = 2,
  alpha = 0.05
)
}
\arguments{
\item{x}{a matrix or data.frame containing the data to plot.}

\item{corr}{type of correlation to perform, set to \code{"spearman"} by
default but can be changed to \code{"pearson"}.}

\item{corr_round}{number of decimal places to round p values and
correlations, set to 2 decimal places by default.}

\item{corr_col_scale}{colour scale to use for the heatmap component of the
correlation plot, set to \code{c("red", "green")} by default.}

\item{order}{column names in the order that they should appear in the plot.}

\item{upper}{type of plots to include in the upper portion of the panel, set
to \code{"scatter"} by default. This can be changed to \code{"heatmap"} to
flip the plot around the other way.}

\item{labels}{vector of labels in an order matching the column names of
\code{x} to replace the feature labels on the histogram panels.}

\item{label_text_size}{size of text in diagonal panels, set to 2 by default.}

\item{label_text_font}{font for text on diagonal panels, set to 1 by default.}

\item{label_text_col}{colour for text on diagonal panels, set to
\code{"black"} by default.}

\item{corr_text_size}{text size for p values and correlation text in heatmap
panels, set to 2 by default.}

\item{corr_text_font}{font for p values and correlation text in heatmap
panels, set to 1 by default.}

\item{corr_text_col}{colour for p values and correlation text in heatmap
panels, set to \code{"black"} by default.}

\item{point_size}{numeric to control the size of points in scatterplots, set
to 1 by default.}

\item{point_col}{colour to use for points in scatterplots, set to
\code{"black"} by default.}

\item{point_type}{type of point to plot in scatterplots, set to 16 default.}

\item{title}{a title for the correlation plot, set to \code{"Correlation
Matrix"} by default.}

\item{title_text_size}{numeric to control the size of the title text, set to
1 by default.}

\item{title_text_col}{colour to use for title text, set to \code{"black"} by
default.}

\item{title_text_font}{font to use for title text, set to 1 by default.}

\item{hist_col}{colour to use for histograms in diagonal panels, set to
\code{"blue"} by default.}

\item{hist_border_col}{colour to use for histogram borders in diagonal
panels, set to \code{"black"} by default.}

\item{hist_border_type}{type of border to use use for histograms in diagonal
panels, set to 1 by default.}

\item{hist_border_width}{width of histogram borders in diagonal panels, set
to 1 by default.}

\item{density_line_col}{colour to use for density distributions in diagonal
panels, set to \code{"black"} by default.}

\item{density_line_type}{type of line to use for density distributions in
diagonal panels, set to 1 by default.}

\item{density_line_width}{width of lines for density distributions in
diagonal panels, set to 1 by default.}

\item{trend_line_type}{type of line to use for trendlines, set to 1 by
default.}

\item{trend_line_col}{colour to use for trendlines, set to \code{"red"} by
default.}

\item{trend_line_width}{width of trendlines, set to 1 by default.}

\item{margins}{vector of length 4 to control margins around each plot, set to
\code{c(0.5, 0.5, 0.5, 0.5)} by default.}

\item{frame}{controls the amount of padding around the outside of the panel,
set to 2 by default.}

\item{alpha}{cutoff point for p value significance, set to \code{0.05} by
default.}
}
\value{
recorded plot
}
\description{
Create a correlation matrix
}
