% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pagel.R
\name{print.pagel}
\alias{print.pagel}
\title{Prints the results of a multi-tree Pagel analysis.}
\usage{
\method{print}{pagel}(pagels)
}
\arguments{
\item{pagels}{Results created by pagel_multi.}
}
\description{
This functions provides print functionality for a multi-tree Pagel analyses results created with pagel_multi.
}
\examples{
print(pagels)
}
