% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_correlate.R
\name{summarize_correlate}
\alias{summarize_correlate}
\title{Creates a summary overview of a correlate analysis.}
\usage{
summarize_correlate(cond = cond)
}
\arguments{
\item{cond}{Results created by get_conditional.}
}
\description{
This functions provides summatized output of a correlate output in tabular format. It will apply pairwise wilcox tests to compare the means of conditional probability distributions for character co-occurence.
}
\examples{
summarize_correlate(cond)
}
